requirements.txt
streamlit
pandas
requests
plotly
google-generativeai